i dont understand...

+++ api/src/main/java/com/api/demo/grid/controller/Account.java
@@ -3,13 +3,21 @@ package com.api.demo.grid.controller;
 
 import com.api.demo.grid.dtos.UserDTO;
 import com.api.demo.grid.exception.ExceptionDetails;
+import com.api.demo.grid.exception.ForbiddenException;
 import com.api.demo.grid.models.User;
 import com.api.demo.grid.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import java.util.Objects;
 
 
 @RestController
@@ -37,7 +45,15 @@ public class Account {
     }
 
     @PostMapping("/remove-user")
-    public void removeUser(@RequestHeader("Authorization") String auth, @RequestParam String username) {
+    public void removeUser(@RequestHeader("Authorization") String auth, @RequestParam String username)
+            throws ForbiddenException {
+
+        // verify if the user requesting is the same as the buyer
+        String usernameRequest = ControllerUtils.getUserFromAuth(auth);
+        if (!Objects.equals(username, usernameRequest) && !mUserService.getUser(usernameRequest).isAdmin()) {
+            throw new ForbiddenException("The user requesting to remove the user account is not an admin and is not " +
+                    "the same as the user requested to remove");
+        }
 
         mUserService.deleteUser(username);
     }
